tests/wpt/tests/css/css-position/parsing/right-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Positioned Layout Module Level 3: parsing right with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-position/#propdef-right">
<meta name="assert" content="right supports only the grammar 'auto | <length-percentage>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("right", "min-content");

test_invalid_value("right", "60");
test_invalid_value("right", "10px 20%");
</script>
</body>
</html>
